listener/requirements.txt
azure-core==1.16.0
azure-eventhub==5.6.0
azure-eventhub-checkpointstoreblob-aio==1.1.4
certifi==2021.5.30
chardet==4.0.0
idna==2.10
requests==2.25.1
six==1.16.0
uamqp==1.4.1
urllib3==1.26.6
python-dotenv==0.18.0
pymongo==3.11.4